Update UI components namespace
