+++ app/View/Components/ui/Modal.php
@@ -1,30 +1,30 @@
 <?php
 
-namespace App\View\Components\UI;
+namespace Components\UI;
 
 use Illuminate\View\Component;
 
-class Dropdown extends Component
+class Modal extends Component
 {
-
 
     /**
      * Create a new component instance.
      *
      * @return void
      */
-    public function __construct($label = null)
+    public function __construct()
     {
         //
     }
 
+
     /**
      * Get the view / contents that represent the component.
      *
-     * @return \Illuminate\Contracts\View\View|\Closure|string
+     * @return \Illuminate\Contracts\View\View|string
      */
     public function render()
     {
-        return view('components.ui.dropdown');
+        return view('components.ui.modal');
     }
 }
