+++ app/View/Components/ui/Select.php
@@ -1,16 +1,15 @@
 <?php
 
-namespace App\View\Components\UI;
+namespace Components\UI;
 
 use Illuminate\View\Component;
 
-class SwitchButton extends Component
+class Select extends Component
 {
     public $key;
 
     public function __construct($key = '')
     {
-
         $this->key = $key;
     }
 
@@ -21,6 +20,6 @@ class SwitchButton extends Component
      */
     public function render()
     {
-        return view('components.ui.switch');
+        return view('components.ui.select');
     }
 }
